Fases funcionando en modo RTC
Hasta el momento se hicieron funcionar las fase en modo de prueba, falta
recibir el dato del I2C y sacar la info del tiempo que estará encendida
cada fase, además se debe crear una función para actualizar el estado de
las fases cada segundo según comando recibido en tiempo real.

+++ Esclava/Drivers/STM32F1xx_HAL_Driver/Src/stm32f1xx_hal_rcc_ex.c
@@ -581,7 +581,7 @@ uint32_t HAL_RCCEx_GetPeriphCLKFreq(uint32_t PeriphClk)
       break;
     }
   }
-  return (uint32_t) frequency;
+  return (frequency);
 }
 
 /**
